Split Player.die() into drop() and die()
Let the drop() function handle the player fall after colliding with a pipe.
Call the die() function when the player is already on the floor and make it handle death.
Make both functions stop movement and animation, to make sure this happens no matter what the player collides with (a pipe or the floor).

Remove the 'player_died' signal from GameManager, there's no need for it after all.